data-pipeline/requirements.txt
apache-beam[gcp]==2.50.0
google-cloud-bigquery==3.11.4
google-cloud-storage==2.10.0
pandas==2.1.1
numpy==1.24.3

